% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transcript_id.R
\name{get_transcript_id}
\alias{get_transcript_id}
\title{Retrieve Available Transcript IDs}
\usage{
get_transcript_id()
}
\value{
A sorted numeric vector of available transcript IDs.
}
\description{
Lists all available transcript IDs based on the \code{.csv} files stored in
the \code{data-raw/transcripts} folder of the \strong{BribeR} package.
}
\details{
Each transcript is named numerically (e.g., \verb{1.csv}, \verb{19.csv}, \verb{104.csv}),
and its ID corresponds directly to that number.
}
\examples{
# Retrieve all available transcript IDs
ids <- get_transcript_id()
ids

# Use those IDs to load specific transcripts
subset <- get_transcripts_raw(n = ids[1:3])

}
\seealso{
\code{\link[=read_transcripts]{read_transcripts()}}, \code{\link[=get_transcripts_raw]{get_transcripts_raw()}}, \code{\link[=get_transcript_speakers]{get_transcript_speakers()}}
}
